api.pl: Prevent sending command parameters
Because we have to include - in the validation of hostnames, this allows
sending parameters like -h or --help via the API.

This is not a security risk as it does not allow space to be used as
valid character in the host paramater.

dig however does not like '--' as parameter separator, it is left
unchanged.

Thank you to maniax for pointing this out.