index.html: Mobile interface improvements
I realized that the interface is very bad in terms of size and
positioning of the elements on mobile, so this commit aims to fix those
issue by significantly improving the readability and usability on
mobile.

This is in no way perfect, but a step in the right direction, without
completely redesigning it for mobile view.